requirements.txt
pymysql
requests
slacker
urllib3
boto3
Pillow

